fix: stop reporting 64-bit integer columns as int32 in profiling
_normalize_type mapped the ibis type int64 to int32, so every 64-bit
integer column was documented as int32 in profiling.md: BigQuery INT64,
Snowflake INTEGER (NUMBER(38,0), which ibis renders as int64), and any
BIGINT column on other backends. The mapping was introduced in #399
without a recorded rationale, and nothing in the codebase depends on the
int32 label — the only consumer is the generated doc read by the agent,
which this actively misleads.

Remove the mapping and update the BigQuery and Snowflake integration
expectations accordingly (the other suites seed 32-bit INTEGER columns
and are unaffected). Add unit tests locking _normalize_type behavior.

+++ cli/tests/nao_core/commands/sync/integration/test_bigquery.py
@@ -147,7 +147,7 @@ def spec(db_config, temp_datasets):
         users_profiling_rows=[
             {
                 "column": "id",
-                "type": "int32",
+                "type": "int64",
                 "total_count": 3,
                 "null_count": 0,
                 "null_percentage": 0.0,
@@ -195,7 +195,7 @@ def spec(db_config, temp_datasets):
         orders_profiling_rows=[
             {
                 "column": "id",
-                "type": "int32",
+                "type": "int64",
                 "total_count": 2,
                 "null_count": 0,
                 "null_percentage": 0.0,
@@ -207,7 +207,7 @@ def spec(db_config, temp_datasets):
             },
             {
                 "column": "user_id",
-                "type": "int32",
+                "type": "int64",
                 "total_count": 2,
                 "null_count": 0,
                 "null_percentage": 0.0,
